Moving SingleSubmitService methods to SubmitService and deleting the class SingleSubmitService
