+++ services/src/main/java/mops/SubmitService.java
@@ -1,7 +1,10 @@
 package mops;
 
+import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 import mops.fragen.Frage;
+import mops.rollen.Student;
 
 @SuppressWarnings("PMD.DataflowAnomalyAnalysis")
 public class SubmitService {
@@ -22,5 +25,21 @@ public class SubmitService {
     if ((antwort != null) && !antwort.equals("")) {
       frage.addAntwort(antwort);
     }
+  }
+
+  public void addStudentAsSubmitted(Fragebogen fragebogen, Student student) {
+    fragebogen.addStudentAsSubmitted(student);
+  }
+
+  public List<Fragebogen> notSubmittedFrageboegen(List<Fragebogen> frageboegen, Student student) {
+    return frageboegen.stream()
+        .filter(fragebogen -> !fragebogen.getAbgegebeneStudierende().contains(student))
+        .collect(Collectors.toList());
+  }
+
+  public List<Fragebogen> frageboegenContaining(List<Fragebogen> frageboegen, String search) {
+    return frageboegen.stream()
+        .filter(fragebogen -> fragebogen.contains(search))
+        .collect(Collectors.toList());
   }
 }
